tests/wpt/web-platform-tests/css/css-size-adjust/parsing/text-size-adjust-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Mobile Text Size Adjustment: parsing text-size-adjust with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-size-adjust/#propdef-text-size-adjust">
<meta name="assert" content="text-size-adjust supports the full grammar 'auto | none | <percentage>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("text-size-adjust", 'auto');
test_valid_value("text-size-adjust", 'none');

test_valid_value("text-size-adjust", '200%');
test_valid_value("text-size-adjust", '100%');
test_valid_value("text-size-adjust", '0%');
</script>
</body>
</html>
